Allowing master manuscript to use the chapters by removing \input commands, commenting out \graphicspath, correcting figure placement to [htb] or [!ht].  Yellownotes and todo forced to be inline.
Gnuplot commands should be called from main directory as "gnuplot --persist gpifile.gpi", so data inclusions must be relative to this directory and not the gfx dir.

Graphics paths removed from \includegraphics so that \graphicspath is the sole organiser of graphics and where to obtain them.

Makefile updated to Chris Monson's 2.2.0-rc11, but I have kept the octave/matlab commands.

Including glossaries and acronyms.

Appendix re-organised.